adjust '#include' style and move macro definition files into folder 'detail'
